requirements.txt
pyte
pywebview
PySide6
qtpy
